move createpaloptions to enum class; palette updates

+++ palmod/Game/GameClass.cpp
@@ -1807,7 +1807,7 @@ BOOL CGameClass::SaveFile(CFile* SaveFile, uint32_t nUnitId)
                     {
                         // Never write the transparency counter.
                         // It's kind of OK to do so since it should be a no-op, but TMNTF is evil and relies upon overlapping palettes.
-                        if ((((nArrayIndex + createPalOptions.nStartingPosition) % createPalOptions.eWriteOutputOptions) != 0) ||
+                        if ((((nArrayIndex + createPalOptions.nStartingPosition) % static_cast<UINT16>(createPalOptions.eWriteOutputOptions)) != 0) ||
                             (nGameFlag == MSHVSF_A)) // We don't support skipping the transparency color for MSHvSF's special override
                         {
                             SaveFile->Write(&m_pppDataBuffer[nUnitCtr][nPalCtr][nArrayIndex], m_nSizeOfColorsInBytes);
